Changes for PR
Seperate JourneyMap plugin code from JourneyMap module code.
Remove import static for non-final fields.
Change Waypoint ID to include dimension and sample name.
IDs were not unique enough and would be identical if samples in
different dimensions were at the same X, Y, Z.
