+++ src/main/java/com/teamacronymcoders/base/modules/journeymap/JourneyMapWaypointListener.java
@@ -1,13 +1,12 @@
 package com.teamacronymcoders.base.modules.journeymap;
 
+import com.teamacronymcoders.base.Base;
 import com.teamacronymcoders.base.event.PlaceWaypointEvent;
 import journeymap.client.api.IClientAPI;
 import journeymap.client.api.display.Waypoint;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 import java.util.Random;
-
-import static com.teamacronymcoders.base.Base.instance;
 
 public class JourneyMapWaypointListener {
     private IClientAPI jAPI;
@@ -21,7 +20,7 @@ public class JourneyMapWaypointListener {
         try {
             jAPI.show(new Waypoint(event.getModId(), event.getId(), event.getName(), event.getDimension(), event.getPos()).setColor(new Random().nextInt(0xffffff)));
         } catch(Throwable t) {
-            instance.getLogger().warning(t.getLocalizedMessage());
+            Base.instance.getLogger().warning(t.getLocalizedMessage());
         }
     }
 }
